fix comparison between int and size_t

+++ htab_clear.c
@@ -5,7 +5,7 @@
 
 void htab_clear(htab_t * t){
     htab_itm_t *previtm = NULL;
-    for(int i=0; i<t->arr_size; i++){
+    for(size_t i=0; i<t->arr_size; i++){
         htab_itm_t *itm = t->arr_ptr[i];
         while(itm!=NULL){
             previtm=itm;
